delete more unused code
